pyproject.toml
[project]
name = "parlanchina"
version = "0.9.0"
description = "Where AI conversations come to play"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask[async]>=3.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "fastmcp>=2.13.1",
    "markdown-it-py>=3.0.0",
    "bleach>=6.1.0",
    "hypercorn>=0.16.0",
    "asgiref>=3.7.0",
    "postgres-mcp>=0.3.0",
    "pywebview>=6.1",
]
